refactor(batchstore): use localstorev2 (#3864)
