+++ pkg/localstorev2/reserve_test.go
@@ -361,7 +361,7 @@ func TestRadiusManager(t *testing.T) {
 
 	t.Run("radius decrease due to under utilization", func(t *testing.T) {
 		t.Parallel()
-		bs := batchstore.New(batchstore.WithReserveState(&postage.ReserveState{Radius: 3}))
+		bs := batchstore.New(batchstore.WithRadius(3))
 
 		storer, err := memStorer(t, dbTestOps(baseAddr, 10, bs, nil, nil, time.Millisecond*50))()
 		if err != nil {
@@ -405,8 +405,8 @@ func TestRadiusManager(t *testing.T) {
 
 	t.Run("radius doesnt change due to non-zero pull rate", func(t *testing.T) {
 		t.Parallel()
-		bs := batchstore.New(batchstore.WithReserveState(&postage.ReserveState{Radius: 3}))
-		storer, err := memStorer(t, dbTestOps(baseAddr, 10, bs, pullsync.NewMockRateReporter(1), nil, time.Second))()
+		bs := batchstore.New(batchstore.WithRadius(3))
+		storer, err := diskStorer(t, dbTestOps(baseAddr, 10, bs, pullsync.NewMockRateReporter(1), nil, time.Millisecond*10))()
 		if err != nil {
 			t.Fatal(err)
 		}
@@ -641,7 +641,7 @@ func TestReserveSampler(t *testing.T) {
 
 		for po := 0; po < maxPO; po++ {
 			for i := 0; i < chunkCountPerPO; i++ {
-				ch := chunk.GenerateTestRandomChunkAt(baseAddr, po).WithBatch(0, 3, 2, false)
+				ch := chunk.GenerateTestRandomChunkAt(baseAddr, po).WithBatch(3, 2, false)
 				// override stamp timestamp to be before the consensus timestamp
 				ch = ch.WithStamp(postagetesting.MustNewStampWithTimestamp(timeVar - 1))
 				chs = append(chs, ch)
@@ -685,7 +685,7 @@ func TestReserveSampler(t *testing.T) {
 		// some of them should definitely make it to the sample based on lex ordering.
 		for po := 0; po < maxPO; po++ {
 			for i := 0; i < chunkCountPerPO; i++ {
-				ch := chunk.GenerateTestRandomChunkAt(baseAddr, po).WithBatch(0, 3, 2, false)
+				ch := chunk.GenerateTestRandomChunkAt(baseAddr, po).WithBatch(3, 2, false)
 				// override stamp timestamp to be after the consensus timestamp
 				ch = ch.WithStamp(postagetesting.MustNewStampWithTimestamp(timeVar + 1))
 				chs = append(chs, ch)
