+++ pkg/postage/batch_test.go
@@ -20,7 +20,7 @@ func TestBatchMarshalling(t *testing.T) {
 	if err != nil {
 		t.Fatal(err)
 	}
-	if len(buf) != 96 {
+	if len(buf) != 95 {
 		t.Fatalf("invalid length for serialised batch. expected 95, got %d", len(buf))
 	}
 	b := &postage.Batch{}
@@ -47,8 +47,5 @@ func TestBatchMarshalling(t *testing.T) {
 	}
 	if a.Immutable != b.Immutable {
 		t.Fatalf("depth mismatch, expected %v, got %v", a.Immutable, b.Immutable)
-	}
-	if a.StorageRadius != b.StorageRadius {
-		t.Fatalf("radius mismatch, expected %v, got %v", a.StorageRadius, b.StorageRadius)
 	}
 }
