+++ pkg/postage/batchstore/mock/store.go
@@ -13,14 +13,13 @@ import (
 	"github.com/ethersphere/bee/pkg/postage"
 	"github.com/ethersphere/bee/pkg/postage/batchstore"
 	"github.com/ethersphere/bee/pkg/storage"
-	"github.com/ethersphere/bee/pkg/swarm"
 )
 
 var _ postage.Storer = (*BatchStore)(nil)
 
 // BatchStore is a mock BatchStorer
 type BatchStore struct {
-	rs                    *postage.ReserveState
+	radius                uint8
 	cs                    *postage.ChainState
 	isWithinStorageRadius bool
 	id                    []byte
@@ -31,8 +30,6 @@ type BatchStore struct {
 	saveErr               error
 	updateErrDelayCnt     int
 	resetCallCount        int
-
-	radiusSetter postage.StorageRadiusSetter
 
 	existsFn func([]byte) (bool, error)
 
@@ -48,7 +45,6 @@ type Option func(*BatchStore)
 func New(opts ...Option) *BatchStore {
 	bs := &BatchStore{}
 	bs.cs = &postage.ChainState{}
-	bs.rs = &postage.ReserveState{}
 	bs.isWithinStorageRadius = true
 	for _, o := range opts {
 		o(bs)
@@ -57,15 +53,9 @@ func New(opts ...Option) *BatchStore {
 }
 
 // WithReserveState will set the initial reservestate in the ChainStore mock.
-func WithReserveState(rs *postage.ReserveState) Option {
+func WithRadius(radius uint8) Option {
 	return func(bs *BatchStore) {
-		bs.rs = rs
-	}
-}
-
-func WithIsWithinStorageRadius(b bool) Option {
-	return func(bs *BatchStore) {
-		bs.isWithinStorageRadius = b
+		bs.radius = radius
 	}
 }
 
@@ -200,49 +190,11 @@ func (bs *BatchStore) PutChainState(cs *postage.ChainState) error {
 	return nil
 }
 
-func (bs *BatchStore) GetReserveState() *postage.ReserveState {
+func (bs *BatchStore) Radius() uint8 {
 	bs.mtx.Lock()
 	defer bs.mtx.Unlock()
 
-	rs := new(postage.ReserveState)
-	if bs.rs != nil {
-		rs.Radius = bs.rs.Radius
-		rs.StorageRadius = bs.rs.StorageRadius
-	}
-	return rs
-}
-
-func (bs *BatchStore) StorageRadius() uint8 {
-	bs.mtx.Lock()
-	defer bs.mtx.Unlock()
-
-	if bs.rs != nil {
-		return bs.rs.StorageRadius
-	}
-	return 0
-}
-
-func (bs *BatchStore) IsWithinStorageRadius(swarm.Address) bool {
-	return bs.isWithinStorageRadius
-}
-
-func (bs *BatchStore) SetStorageRadiusSetter(r postage.StorageRadiusSetter) {
-	bs.radiusSetter = r
-}
-
-func (bs *BatchStore) SetStorageRadius(f func(uint8) uint8) error {
-	bs.mtx.Lock()
-	defer bs.mtx.Unlock()
-
-	bs.rs.StorageRadius = f(bs.rs.StorageRadius)
-	if bs.radiusSetter != nil {
-		bs.radiusSetter.SetStorageRadius(bs.rs.StorageRadius)
-	}
-	return nil
-}
-
-func (bs *BatchStore) Unreserve(_ postage.UnreserveIteratorFn) error {
-	panic("not implemented")
+	return bs.radius
 }
 
 // Exists reports whether batch referenced by the give id exists.
