+++ pkg/postage/stamp.go
@@ -161,7 +161,7 @@ func ValidStamp(batchStore Storer) ValidStampFn {
 		if err = stamp.Valid(chunk.Address(), b.Owner, b.Depth, b.BucketDepth, b.Immutable); err != nil {
 			return nil, err
 		}
-		return chunk.WithStamp(stamp).WithBatch(b.StorageRadius, b.Depth, b.BucketDepth, b.Immutable), nil
+		return chunk.WithStamp(stamp).WithBatch(0, b.Depth, b.BucketDepth, b.Immutable), nil // TODO: remove radius arg
 	}
 }
