+++ pkg/postage/testing/batch.go
@@ -121,7 +121,4 @@ func CompareBatches(t *testing.T, want, got *postage.Batch) {
 	if want.Depth != got.Depth {
 		t.Fatalf("depth: want %v, got %v", want.Depth, got.Depth)
 	}
-	if want.StorageRadius != got.StorageRadius {
-		t.Fatalf("storageRadius: want %v, got %v", want.StorageRadius, got.StorageRadius)
-	}
 }
